+++ core/templates/pages/contributor-dashboard-admin-page/contributor-dashboard-admin-page.module.ts
@@ -18,7 +18,7 @@
 
 import { APP_INITIALIZER, NgModule, StaticProvider } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { BrowserModule } from '@angular/platform-browser';
+import { BrowserModule, HAMMER_GESTURE_CONFIG } from '@angular/platform-browser';
 import { downgradeComponent } from '@angular/upgrade/static';
 import { HttpClientModule } from '@angular/common/http';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
@@ -29,6 +29,7 @@ import { ContributorDashboardAdminNavbarComponent } from './navbar/contributor-d
 import { platformFeatureInitFactory, PlatformFeatureService } from 'services/platform-feature.service';
 import { HybridRouterModuleProvider } from 'hybrid-router-module-provider';
 import { ToastrModule } from 'ngx-toastr';
+import { MyHammerConfig, toastrConfig } from 'pages/oppia-root/app.module';
 
 @NgModule({
   imports: [
@@ -59,6 +60,10 @@ import { ToastrModule } from 'ngx-toastr';
       useFactory: platformFeatureInitFactory,
       deps: [PlatformFeatureService],
       multi: true
+    },
+    {
+      provide: HAMMER_GESTURE_CONFIG,
+      useClass: MyHammerConfig
     }
   ]
 })
@@ -69,7 +74,6 @@ class ContributorDashboardAdminPageModule {
 
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { downgradeModule } from '@angular/upgrade/static';
-import { toastrConfig } from 'pages/oppia-root/app.module';
 
 const bootstrapFnAsync = async(extraProviders: StaticProvider[]) => {
   const platformRef = platformBrowserDynamic(extraProviders);
