+++ core/templates/pages/error-pages/error-iframed-page/error-iframed-page.module.ts
@@ -17,7 +17,7 @@
  */
 
 import { APP_INITIALIZER, NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
+import { BrowserModule, HAMMER_GESTURE_CONFIG } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
@@ -28,7 +28,7 @@ import { platformFeatureInitFactory, PlatformFeatureService } from
 import { ErrorIframedPageRootComponent } from './error-iframed-root.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ToastrModule } from 'ngx-toastr';
-import { toastrConfig } from 'pages/oppia-root/app.module';
+import { MyHammerConfig, toastrConfig } from 'pages/oppia-root/app.module';
 
 
 @NgModule({
@@ -56,6 +56,10 @@ import { toastrConfig } from 'pages/oppia-root/app.module';
       useFactory: platformFeatureInitFactory,
       deps: [PlatformFeatureService],
       multi: true
+    },
+    {
+      provide: HAMMER_GESTURE_CONFIG,
+      useClass: MyHammerConfig
     }
   ],
   bootstrap: [ErrorIframedPageRootComponent]
